[refactor] 경로 이동 및 파일 명 변경
